[feat] Tile variations.

+++ libs/rustaria-rendering/src/chunk_drawer/chunk.rs
@@ -129,8 +129,8 @@ impl BakedChunk {
                         let matrix = &tile_neighbor_row[x];
                         drawer.push(
                             builder,
-                            (pos.x as f32 * CHUNK_SIZE as f32) + (x as f32),
-                            (pos.y as f32 * CHUNK_SIZE as f32) + (y as f32),
+                            (pos.x * CHUNK_SIZE as u32) + (x as u32),
+                            (pos.y * CHUNK_SIZE as u32) + (y as u32),
                             super::tile::TileConnectionKind::new(
                                 matrix.up,
                                 matrix.down,
